WIP: first attempt at tree-shaking top-level class instantiation with constructors
This is a follow up on #676.

It's not quite clear to me how to wire up the scopes so that 'this' is found at the right time.

This seems to work, but there are some ugly things I had to do to get the test to pass.

Please take a look and suggest changes.

Fixes #674