Generate an interop default block for external modules on iife/umd
Currently, when options.interop == true (default) and an external
module is imported, cjs generates an interop block that looks
like this:

```js
  function _interopDefault (ex) { return (ex && (typeof ex === 'object')
&& 'default' in ex) ? ex['default'] : ex; }
```

for iife and umd, there's a simliar block generated by `getInteropBlock()`
that looks like this:

```js
  var foo__default = 'default' in foo ? foo['default'] : foo;
```

but for external modules that export as a namespace, it's generated as:

```js
  var foo__default = ${name}['default'];
```
(I don't know why this is more appropriate for namespace exports, but
it's not the issue I'm facing)

Now, when an external module is traced (ExternalModule.traceExports)
to have both namespace and named imports, the namespace case takes
priority.

Changed to generate the latter expression whenever named imports are
found, regardless of namespace imports.