Show errors for every build in multiconfig
Currently if one build fails then the next one is swallowed.

I'm working on size-snapshot plugin which fails if size snapshot is
not matched. The problem is that with umd, esm and cjs configs we will
see only umd error which may confuse why there is not error for another
builds.

My solution is to lift `process.exit(1)`.