Don't test entry points for external
Ref https://github.com/rollup/rollup/issues/1343

I just found that external is almost the same as `resolveId`. Currently
for the stuff like ignoring all `module` paths we should consider entry
point which may look similar and do not start with '.' or '/'. This is
annoying since external entry points are not allowed.

In this diff I suggest to not test entry points for external at all and
skip this check in user configs.