Show how to skip imports for optional plugins
The most common Rollup plugins are pretty fast to import (by Node
standards), but even those will commonly add on a handful of
milliseconds, and it’s not unheard of to have imports literally take
seconds. If you don’t need it, why load it? I think that this is a good
pattern to recommend.

eslint and stylelint are two good examples of unreasonably slow modules
that will commonly be optional. stylelint can take a second or two, and
eslint a few hundred milliseconds, because neither of them care enough
about startup time. https://github.com/stylelint/stylelint/issues/2454
was me reporting it for stylelint. I wish more people used Rollup on
their packages before submitting them to npm.