Enable strictNullChecks
There were 400 errors to fix and most of the mwere fixed by adding explicit type
assertions at every error site. It would be a good idea to go back and
clean some of these up later.

There are a few places that I changed the logic or the types, which I've
tried to point out in the pull request comments.