Ensure --watch mode exits correctly when stdin is closed.
This is necessary when rollup is used as the front end for an Elixir
Phoenix application. In dev mode, the phoenix watchers run the rollup
command as a child process and the underlying erlang port mechanism
assumes that the child process closes when the parent stdin is
closed. Without this fix the old zombie processes never close and
accumulate as you stop and start in development mode.

This was discussed and fixed in 2017 in the original rollup-watcher
repository:

  * https://github.com/rollup/rollup-watch/issues/30#issuecomment-316468552
  * https://github.com/rollup/rollup-watch/pull/57/files

It appears to then have been removed as a side effect of another change

  * https://github.com/rollup/rollup/pull/2410

Finally, was requested back in a PR that never got merged because of
some doubt around the need to hide it behind a CLI flag

  * https://github.com/rollup/rollup/pull/2653/files

For reference, webpack implements this hidden behind the --watch-stdin
CLI flag.