Upgrade chokidar and fsevents for Apple M1 compatibility (#3945)
This PR updates `chokidar` and `fsevents` in package.json.

`fsevents` ~2.1.2 was deprecated, and mentions to use 2.2 or 2.3 instead. This is the message you see when installing rollup as a dependency: `rollup > fsevents@2.1.3: "Please update to latest v2.3 or v2.2"`. The [commits](https://github.com/fsevents/fsevents/commits/master) there imply it's for Apple M1 compatibility.

`chokidar` also updated to that version of `fsevents`, so I updated that here too.

If you see package-lock.json, you'll see the older version of 2.1.3 got added to mocha's dependencies. They use the older version, and haven't released a new version yet.