Rename annotations in acron.Node
When parsing the JS code we add attributes to acorn.Node for later use:
* annotations: acorn.Commnet - the comment that defined the annotation
* annotatedPure: boolean - declares this node as pure call/new expression

This commit puts all these attributes under one name: `_rollupAnnotations` with
a strong type.

Thus we clarify that these attributes are for internal use only and decrease
the change of ever coliding with internal/future attributes of `acorn.Node`.