Introduce `sourcemapIgnoreList` predicate.
This adds a new `sourcemapIgnoreList` config predicate that works
similar to the existing `sourcemapPathTransform`, and can be provided to
mark certain source files as ignore-listed when generating source maps.

This is useful for libraries and frameworks that provide code that is
not relevant to the debugging experience of the customer. Chrome
DevTools already supports this new `x_google_ignoreList` field in the
source maps[^1].

Doc: https://goo.gle/devtools-ignoreList-adoption

[^1]: https://developer.chrome.com/blog/devtools-better-angular-debugging