Allow to pass "true" to InputOptions.cache
Vite has changed the default of this option to "false" (https://github.com/vitejs/vite/pull/11454), therefore it is needed to be able to set it to "true" explicitly. This may be useful in other situations too, e.g. extending some defaults options object which contains this option set to "false".