Allow to pass "true" to InputOptions.cache (#4859)
* Allow to pass "true" to InputOptions.cache

Vite has changed the default of this option to "false" (https://github.com/vitejs/vite/pull/11454), therefore it is needed to be able to set it to "true" explicitly. This may be useful in other situations too, e.g. extending some defaults options object which contains this option set to "false".

* adapt documentation of "cache" option

* fix type error

* Fix linting

---------

Co-authored-by: Lukas Taegert-Atkinson <lukastaegert@users.noreply.github.com>
Co-authored-by: Lukas Taegert-Atkinson <lukas.taegert-atkinson@tngtech.com>