feat: mark files in a `node_modules` as ignore-listed by default
Following up on the addition of the `sourcemapIgnoreList` output option,
this adds a default which marks all files within a `node_modules` folder
as ignore-listed. This way most projects using rollup will need no
additional configuration regarding the ignore-listing and automatically
benefit.

This also allows to specify `false` for the `sourcemapIgnoreList` output
option, which completely disables the generation of an ignore-list.

Doc: https://goo.gle/devtools-ignoreList-adoption