perf(sourcemaps): re-encode sourcemaps to reduce memory
With the previous commit, we defer decoding the mappings field until we
know they are going to be used. This change aims to further reduce peak
heap usage by allowing for the decoded arrays to be cleaned up sooner
instead of retained until the end of the build.

When we are likely to be done with the `originalSourcemap` and
`sourcemapChain`, we can re-encode them which allows for the larger
decoded array to be garbage collected while retaining the smaller VLQ
encoded strings instead. This commit also makes a change to the previous
commit's caching behavior in that the original encoded mapping is no
longer being held onto once we have a decoded array (this is
closer to the original behavior pr and reduces memory usage further).
This does mean, however, that we need need to recalculate the encoded
string from the decoded arrays instead of just dropping the decoded
array reference. Somewhat counter-intuitively, I found re-encoding to be
both faster (2%) and use less memory than the variant that kept the
original encoded string around.

There are two places where we are currently reverting from decoded ->
encoded mappings:
1) At module.setSource which is passed sourcemaps from transform or from
the rollup cache
2) As part of collapse/folding the module sourcemaps during renderChunk

For the transform case, that operation will frequently be a noop since
the decoding is now lazy and would not have been invoked unless a plugin
called `getCombinedSourcemap()` explicitly. If that were the case, the
combined sourcemap would have its decoded mappings thrown away before
the sourcemaps are stored on the module resulting in them potentially
being decoded a second time when the chunk is rendered.

The second case, renderChunk, appears to make up the majority of the
memory usage reduction for me. It is often times during rendering chunks
that I hit OOM errors when building with lower `--max-old-space-size`
settings. By aggressively removing the cached decoded mappings arrays
when we are done with them, it gives more headroom for the build to
succeed. This effect is particularly noticable when working with large
sourcemaps.