chore: drop ExistingDecodedSourceMap.mappings setter
Removing the setter for the mappings field reduces complexity of the
cache behavior somewhat, and increases test coverage of the feature.

We expect the cases where someone might be modifying mappings on an
existing sourcemap to be rare. I think a user would need to hook into
the cache and modify it there. For anyone doing that, this would now
fail because the property is not writeableso I have modified the type to
reflect this readonly nature.
The workaround for any users that encounter problems here is to just
replace the entire sourcemap.

```ts
module.originalSourceMap = {
  ...module.originalSourceMap,
  mappings: myNewMappings
};
```