perf(sourcemap): lazy compute decoded mappings (#5075)
* perf(sourcemap): lazy compute decoded mappings

By deferring the decoding of the sourcemap mappings field, we can avoid
extra computation and garbage generation, both of which appear to save
a significant amount of time.

In my test application (using vite), I saw a reduction in total build
time of about 28.4%, and a reduction in peak heap size of 19.6% (706 MB)

Project Info:
* Node: v18.17.0
* Vite v4.4.7
* Rollup v3.26.3
* Module Count: 5490 modules transformed
* Chunk Count: 54 chunks rendered

Before:
* Average vite build time (5 runs): 30.89 seconds
* Average peak heap size (5 runs): 3594.83 MB
* Time spent in GC: ~13.7 seconds
* Time spent in `decode()`: 3.11 seconds

After:
* Average vite build time (5 runs): 22.13 seconds
* Average peak heap size (5 runs): 2888.2 MB
* Time spent in GC: ~5.9 seconds
* Time spent in `decode()`: 1.70 seconds

* perf(sourcemaps): re-encode sourcemaps to reduce memory

With the previous commit, we defer decoding the mappings field until we
know they are going to be used. This change aims to further reduce peak
heap usage by allowing for the decoded arrays to be cleaned up sooner
instead of retained until the end of the build.

When we are likely to be done with the `originalSourcemap` and
`sourcemapChain`, we can re-encode them which allows for the larger
decoded array to be garbage collected while retaining the smaller VLQ
encoded strings instead. This commit also makes a change to the previous
commit's caching behavior in that the original encoded mapping is no
longer being held onto once we have a decoded array (this is
closer to the original behavior pr and reduces memory usage further).
This does mean, however, that we need need to recalculate the encoded
string from the decoded arrays instead of just dropping the decoded
array reference. Somewhat counter-intuitively, I found re-encoding to be
both faster (2%) and use less memory than the variant that kept the
original encoded string around.

There are two places where we are currently reverting from decoded ->
encoded mappings:
1) At module.setSource which is passed sourcemaps from transform or from
the rollup cache
2) As part of collapse/folding the module sourcemaps during renderChunk

For the transform case, that operation will frequently be a noop since
the decoding is now lazy and would not have been invoked unless a plugin
called `getCombinedSourcemap()` explicitly. If that were the case, the
combined sourcemap would have its decoded mappings thrown away before
the sourcemaps are stored on the module resulting in them potentially
being decoded a second time when the chunk is rendered.

The second case, renderChunk, appears to make up the majority of the
memory usage reduction for me. It is often times during rendering chunks
that I hit OOM errors when building with lower `--max-old-space-size`
settings. By aggressively removing the cached decoded mappings arrays
when we are done with them, it gives more headroom for the build to
succeed. This effect is particularly noticable when working with large
sourcemaps.

* chore: drop ExistingDecodedSourceMap.mappings setter

Removing the setter for the mappings field reduces complexity of the
cache behavior somewhat, and increases test coverage of the feature.

We expect the cases where someone might be modifying mappings on an
existing sourcemap to be rare. I think a user would need to hook into
the cache and modify it there. For anyone doing that, this would now
fail because the property is not writeableso I have modified the type to
reflect this readonly nature.
The workaround for any users that encounter problems here is to just
replace the entire sourcemap.

```ts
module.originalSourceMap = {
  ...module.originalSourceMap,
  mappings: myNewMappings
};
```

---------

Co-authored-by: Adam Hines <ahines@factset.com>
Co-authored-by: Lukas Taegert-Atkinson <lukastaegert@users.noreply.github.com>