perf: reducing ast node memory overhead (#5133)
* perf: pack booleans in bitfield

Booleans are represented by numbers and take up 4 bytes each at the time
of this commit in v8. This commit packs all the boolean fields of the
AST nodes into a single 32-bit integer which ammortizes the cost of the
boolean fields. The end result is smaller AST nodes.

* chore: use bit shifts instead of binary literals

* chore: remove context wrapper accessor from node

We now just access through `this.scope.context` directly.

* chore: cleaning up keys assignment

---------

Co-authored-by: Lukas Taegert-Atkinson <lukastaegert@users.noreply.github.com>