Make the path a required parameter for includeVariableInModule
This will result in including the given path. However, if a
variable is later called with different paths, we also want
to include those paths. Thus if the variable is included, we
still call includePath on the variable and rely on the included
path tracking logic of the variable to prevent infinite cycles.