Include context with call arguments
We will need this as calling a function on an object, we need to make sure all
paths on the object accessed by the function are included. At the moment, this
includes many more paths as the logic is not yet very refined.
In the future, this should probably become includeInteractionArgumentsAtPath