Include "this"-properties on member expressions like call arguments
This will integrate context handling into the interaction logic we have. It
will also allow to extend the argument inclusion logic to methods called on
objects, otherwise we would have needed to include too many additional things.

The intention here is that in the future, this would also allow us to implement
safe support for function.call and function.apply.